Commit 1.5.23

+++ database/migrations/2023_04_30_062535_create_user_activity_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('user_activity', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('game_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('game_id');
             $table->unsignedFloat('training_time');
             $table->timestamps();
